Create checkfornewrun.bash
The script should be run every ten minutes on the preprocessing server to start preprocessing when the run is complete.

First it makes sure that preprocessing has not already been started. Then it makes sure that we not already have a lot of demultiplexing processing running. If not it will start the preprocessing.